pyproject.toml
[project]
name = "survey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairocffi>=1.7.1",
    "igraph>=0.11.8",
    "ipykernel>=6.29.5",
    "ipython>=9.0.2",
    "jupyter>=1.1.1",
    "jupytext>=1.16.7",
    "matplotlib>=3.10.1",
    "mpmath>=1.3.0",
    "networkx>=3.4.2",
    "neuralrecommender>=0.1.0",
    "nmf-torch>=0.1.1",
    "notebook>=7.4.4",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyarrow>=21.0.0",
    "pycairo>=1.27.0",
    "pyqt5>=5.15.11",
    "scienceplots>=2.1.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "setuptools>58.0.0",
    "tk>=0.1.0",
    "torch>=2.6.0",
    "torch-scatter>=2.1.2",
    "tqdm>=4.67.1",
    "wheel>=0.45.1",
]
